% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oag_fetch.R
\name{available_oag_tokens}
\alias{available_oag_tokens}
\title{Return how many request tokens are available to query the OpenAIRE Graph API}
\usage{
available_oag_tokens(realm = "openaire")
}
\arguments{
\item{realm}{A string used to identify the "realm"}
}
\value{
The number of request tokens available to query the OpenAIRE Graph
API.
}
\description{
Return how many request tokens are available to query the OpenAIRE Graph API
}
\keyword{internal}
