% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{concat_and_or}
\alias{concat_and_or}
\title{Concatenate string arguments with an "AND" or "OR" as separator}
\usage{
concat_and_or(..., operator)
}
\arguments{
\item{...}{The string(s) to negate or to concatenate with "AND"/"OR". All
values passed to the \code{...} argument must be bare strings. In \code{concat_not()},
only a single string can be passed to \code{...}.}

\item{operator}{A string indicating the logical operator to use to
concatenate the arguments passed to \code{...}. Only "AND" and "OR" (case
sensitive) are accepted.}
}
\value{
A string of class \code{oag_str}.
}
\description{
In the OpenAIRE Graph API, logical operators can be used in filters. This
function allows concatenating series of strings with the logical operators
"AND"/"OR", following the API documentation.
}
\keyword{internal}
