% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oag_options.R
\name{fmt_opt_cursor}
\alias{fmt_opt_cursor}
\title{Check and format the "cursor" option passed to \code{oag_set_options()}}
\usage{
fmt_opt_cursor(cursor, is_cursor_next = FALSE, call)
}
\arguments{
\item{cursor}{A logical indicating whether to use cursor-based paging when
querying the OpenAIRE Graph API (\code{TRUE}) or not (\code{FALSE}). Note that the
\code{cursor} (cursor-based paging) and \code{page} (offset-based paging) arguments
cannot be used simultaneously.}

\item{is_cursor_next}{\strong{Do not use this argument}. It is used internally by
\code{oag_fetch()} to indicate whether the \code{cursor} argument must be interpreted
as the method to use ("cursor-based paging") or as the next cursor.}

\item{call}{The caller environment passed to \code{cli::cli_abort()}.}
}
\value{
A string following OpenAIRE Graph API documentation that can be
added to an API call to specify that cursor-based paging should be used. If
\code{cursor} is NULL at the input, the result returned will also be NULL.
}
\description{
Check and format the "cursor" option passed to \code{oag_set_options()}
}
\keyword{internal}
