% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oag_options.R
\name{fmt_opt_sorting}
\alias{fmt_opt_sorting}
\title{Check and format sorting option passed to \code{oag_set_options()}}
\usage{
fmt_opt_sorting(sortBy, entity, call)
}
\arguments{
\item{sortBy}{A named character vector with the field(s) on which to sort the
results. Names must be valid sorting fields, only accepting the values "ASC"
(ascending) and "DESC" (descending) as input. Note that the fields on which
results can be sorted vary across entities. To get the list of sorting fields
for a given entity use \code{get_sorting_fields()}.}

\item{call}{The caller environment passed to \code{cli::cli_abort()}.}
}
\value{
A string following OpenAIRE Graph API documentation that can be
added to an API call to control sorting. If \code{sortBy} is NULL at the input,
the result returned will also be NULL.
}
\description{
Check and format sorting option passed to \code{oag_set_options()}
}
\keyword{internal}
