% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oag_options.R
\name{fmt_opt_stats}
\alias{fmt_opt_stats}
\title{Check and format the "includeStats" option passed to \code{oag_set_options()}}
\usage{
fmt_opt_stats(includeStats, .call)
}
\arguments{
\item{includeStats}{A scalar logical indicating whether statistics about the
query should be included (\code{TRUE}) or not (\code{FALSE}) in the results.}

\item{.call}{The caller environment passed to \code{cli::cli_abort()}.}
}
\value{
A string following OpenAIRE Graph API documentation that can be
added to an API call to control the inclusion of query statistics. If
\code{includeStats} is NULL at the input, the result returned will also be NULL.
}
\description{
Check and format the "includeStats" option passed to \code{oag_set_options()}
}
\keyword{internal}
