% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oag_fetch.R
\name{get_oag_api_url}
\alias{get_oag_api_url}
\title{Get the OpenAIRE Graph API url for a given entity}
\usage{
get_oag_api_url(entity)
}
\arguments{
\item{entity}{A string with the entity to query. It must match an entity
returned by \code{oag_entities()}.}
}
\value{
A string with the URL to the API for the given entity.
}
\description{
Get the OpenAIRE Graph API url for a given entity
}
\examples{
get_oag_api_url("datasources")
}
