% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oag_options.R
\name{get_sorting_fields}
\alias{get_sorting_fields}
\title{Get the list of available fields for sorting in the entities}
\usage{
get_sorting_fields(entity)
}
\arguments{
\item{entity}{A string with the entity to query. It must match an entity
returned by \code{oag_entities()}.}
}
\value{
A character vector with the fields on which query results for the
given entity can be sorted on.
}
\description{
Get the list of available fields for sorting in the entities
}
\examples{
get_sorting_fields("projects")
}
