% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oag_token.R
\name{oag_api_token}
\alias{oag_api_token}
\title{Get the token to connect to the OpenAIRE Graph API}
\usage{
oag_api_token(
  refresh_token = Sys.getenv("oag_api_refresh_token"),
  token = Sys.getenv("oag_api_token")
)
}
\arguments{
\item{refresh_token}{A string with a refresh token. The default is to check
the environment variables for a variable called \code{oag_api_refresh_token}.}

\item{token}{A string with a regular token. The default is to check the
environment variables for a variable called \code{oag_api_token}.}
}
\value{
A token to access the OpenAIRE Graph API or NULL.
}
\description{
The OpenAIRE Graph API can be accessed via authenticated and non
authenticated requests. One method for authenticated access is to use a
personal access token (PAT). Two types of tokens exists:
\itemize{
\item regular token: valid for one hour and must be regenerated at the end
of the validity period.
\item refresh token: valid for one month, it allows to programmatically
refresh an access token for one hour during the whole validity period.
}
}
\details{
By default, the function will check for the existence of a refresh token in
the environment variables. If there is one, it will use it to create an
access token, valid for one hour, based on the refresh token. After that, a
new access token will be refreshed when the 1-hour period is over.

If no refresh token exist (or if the user requested not to use it), the
function will fall back to using the regular token stored in the environment
variables. In the end, if there are no tokens in the environment variables,
the function will return nothing.
}
\examples{
\dontrun{
# Default is to use the refresh token stored in the environment variable
# `oag_api_refresh_token`.
token <- oag_api_token()
# To force the function using the regular token
token <- oag_api_token(refresh_token = NULL)
}
}
