% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oag_fetch.R
\name{oag_entities}
\alias{oag_entities}
\title{Get the list of entities available in the OpenAIRE Graph API}
\usage{
oag_entities()
}
\value{
A character vector with the OpenAIRE graph entities.
}
\description{
The entities available are:
\itemize{
\item Research products (publications, datasets, software, and other
research products)
\item Organizations (universities, research organizations, funders, and
companies)
\item Data sources (repositories, journals, aggregators, and funder
databases whose content is available in the OpenAIRE Graph)
\item Projects (funded research project grants)
\item Persons (authors and contributors involved in research products)
}
The list of entities exposed in the API is available at
\url{https://graph.openaire.eu/docs/apis/graph-api/overview}.
}
\examples{
oag_entities()
}
