% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oag_fetch.R
\name{oag_fetch}
\alias{oag_fetch}
\title{Perform a request and fetch the response from OpenAIRE Graph API}
\usage{
oag_fetch(query)
}
\arguments{
\item{query}{An \code{oag_query} object formatted with \code{oag_query()} or a bare
string with a valid query to the API.}
}
\value{
The results from the performed query in a JSON format.
}
\description{
Based on an \code{oag_query} object or a string passed to the function, it
performs a request to the OpenAIRE Graph API and fetch the response back.
Note that the function use throttling to make sure that the performed
requests never exceed the rate limit (60 per hour for unauthenticated
requests -- authenticated request is not implemented yet). To see the
remaining number of requests and the time to wait when the rate has been
exceeded, run \code{httr2::throttle_status()} and look for the row where \code{realm}
is "api.openaire.eu".
}
\examples{
\dontrun{
# Create a query then pipe it to `oag_fetch()` to get the results
rp_data <- oag_query(
  "research-products",
  publicationYear = "2020",
  options = oag_options(sortBy = c(relevance = "ASC"))
) |>
  oag_fetch()
}
}
