% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oag_fetch.R
\name{oag_query}
\alias{oag_query}
\alias{oag_fetch}
\title{Compose a query and fetch the data from the OpenAIRE Graph API}
\usage{
oag_query(entity, ..., options = NULL)

oag_fetch(entity, ..., options = NULL, token = oag_api_token())
}
\arguments{
\item{entity}{A string with the entity to query. It must match an entity
returned by \code{oag_entities()}.}

\item{...}{Filter(s) to use when composing the query. The filter(s) must be
named and be a valid filter as referenced in the OpenAIRE Graph API
documentation. For example, \code{publicationYear = "2020"} can be used to
filter publication from 2020 when working with the "research-products"
entity.}

\item{options}{A set of options (created with \code{oag_options()}) used to
compose the query. Options allow to control sorting, paging, etc. See
\code{?oag_options()} for more details.}

\item{token}{A valid regular or refresh token to authenticate to the
OpenAIRE Graph API, granting the user of 7'200 requests per hour. If no
token is passed, the request will be unauthenticated, with a rate limit of
60 requests per hour.}
}
\value{
A string with the query URL for \code{oag_query()} and he results from
the performed query in a JSON format for \code{oag_fetch()}.
}
\description{
The OpenAIRE Graph API works with filters and options (see
\url{https://graph.openaire.eu/docs/apis/graph-api/}). The functions
interpret additional (named) argument(s) passed to the \code{...} argument as
filter(s). The filter(s) must be named with the name of the field to filter,
as described in the OpenAIRE Graph API. Options can be specified with
\code{oag_options()}.

Note that \code{oag_fetch()} is based on \code{oag_request()}, which use throttling
to make sure that the performed requests never exceed the rate limit (60 per
hour for unauthenticated requests, compared to 7200 for authenticated ones).
}
\examples{
# Create a query
oag_query(
  "research-products",
  publicationYear = "2020",
  options = oag_options(sortBy = c(relevance = "ASC"))
)
\dontrun{
# Create a query then pipe it to `oag_fetch()` to get the results
rp_data <- oag_query(
  "research-products",
  publicationYear = "2020",
  options = oag_options(sortBy = c(relevance = "ASC"))
) |>
  oag_request()
}
}
