% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oag_fetch.R
\name{oag_request}
\alias{oag_request}
\title{Perform the request to the OpenAIRE Graph API}
\usage{
oag_request(url, token = oag_api_token())
}
\arguments{
\item{url}{An \code{oag_query} object or a string with a valid URL to make a call
to the API.}

\item{token}{A string with a valid token to call the API. \code{oag_api_token()}
automatically handle retrieving the user's token when saved as environment
variable. See \code{?oag_api_token()} for more details.}
}
\value{
The response from the API as a JSON object.
}
\description{
Perform the request to the OpenAIRE Graph API
}
\examples{
\dontrun{
oag_query(
  "research-products",
  publicationYear = "2020",
  options = oag_options(sortBy = c(relevance = "ASC"))
) |>
  oag_request()
}
}
