% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{concat_and}
\alias{concat_and}
\alias{concat_or}
\alias{concat_not}
\title{Concatenate strings with a logical operator (AND-OR-NOT)}
\usage{
concat_and(...)

concat_or(...)

concat_not(...)
}
\arguments{
\item{...}{The string(s) to negate or to concatenate with "AND"/"OR". All
values passed to the \code{...} argument must be bare strings. In \code{concat_not()},
only a single string can be passed to \code{...}.}
}
\value{
A string of class \code{oag_str}.
}
\description{
In the OpenAIRE Graph API, logical operators can be used in filters. These
functions allows concatenating series of strings with "AND"/"OR" operator, or
to negate a string with the "NOT" operator, following the API documentation.
}
\examples{
concat_and("science", "history")
concat_or("biology", "humanities", "geology")
concat_not("model")

# Combining different "concat" functions
concat_and("science", "history", concat_not("model"))
concat_and(concat_or("biology", "humanities"), concat_or("history", "geology"))

# In a query context
oag_query("datasources", search = concat_or("biology", "humanities", "geology"))
}
