% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oag_options.R
\name{fmt_opt_page}
\alias{fmt_opt_page}
\title{Check and format the "page" option passed to \code{oag_set_options()}}
\usage{
fmt_opt_page(page, call)
}
\arguments{
\item{page}{A positive integer indicating the page the query should return.
Using this argument trigger an offset-based paging query. Note that the
\code{page} (offset-based paging) and \code{cursor} (cursor-based paging) arguments
cannot be used simultaneously.}

\item{call}{The caller environment passed to \code{cli::cli_abort()}.}
}
\value{
A string following OpenAIRE Graph API documentation that can be
added to an API call to specify the page to return from a query. If \code{page} is
NULL at the input, the result returned will also be NULL.
}
\description{
Check and format the "page" option passed to \code{oag_set_options()}
}
\keyword{internal}
