% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oag_options.R
\name{fmt_opt_page_size}
\alias{fmt_opt_page_size}
\title{Check and format the "pageSize" option passed to \code{oag_set_options()}}
\usage{
fmt_opt_page_size(pageSize, .call)
}
\arguments{
\item{pageSize}{A positive integer indicating the number of records per page
a query should return. The OpenAIRE Graph API has a limit of 100 records per
page.}

\item{.call}{The caller environment passed to \code{cli::cli_abort()}.}
}
\value{
A string following OpenAIRE Graph API documentation that can be
added to an API call to specify the page size of a query. If \code{pageSize} is
NULL at the input, the result returned will also be NULL.
}
\description{
Check and format the "pageSize" option passed to \code{oag_set_options()}
}
\keyword{internal}
