% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oag_parser.R
\name{init_res_prod_df}
\alias{init_res_prod_df}
\alias{init_data_sources_df}
\alias{init_orgs_df}
\alias{init_proj_df}
\title{Initiate an empty tibble to used to store parsed entities objects}
\usage{
init_res_prod_df(
  type = c("publication", "dataset", "software", "other"),
  selection = NULL
)

init_data_sources_df(selection = NULL)

init_orgs_df(selection = NULL)

init_proj_df(selection = NULL)
}
\arguments{
\item{type}{A string with the type of research products to extract}

\item{selection}{A character vector with the variables of the research
products object that should be returned.}
}
\value{
An empty tibble.
}
\description{
The function generate an empty tibble that will be used to collect the parsed
data coming from a query to an OpenAIRE Graph entity.
Graph API.
}
\keyword{internal}
