% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oag_options.R
\name{oag_options}
\alias{oag_options}
\title{Collect the options to use in a query to the OpenAIRE Graph API}
\usage{
oag_options(
  sortBy = NULL,
  includeStats = NULL,
  pageSize = NULL,
  page = NULL,
  cursor = NULL
)
}
\arguments{
\item{sortBy}{A named character vector with the field(s) on which to sort the
results. Names must be valid sorting fields, only accepting the values "ASC"
(ascending) and "DESC" (descending) as input. Note that the fields on which
results can be sorted vary across entities. To get the list of sorting fields
for a given entity use \code{get_sorting_fields()}.}

\item{includeStats}{A scalar logical indicating whether statistics about the
query should be included (\code{TRUE}) or not (\code{FALSE}) in the results.}

\item{pageSize}{A positive integer indicating the number of records per page
a query should return. The OpenAIRE Graph API has a limit of 100 records per
page.}

\item{page}{A positive integer indicating the page the query should return.
Using this argument trigger an offset-based paging query. Note that the
\code{page} (offset-based paging) and \code{cursor} (cursor-based paging) arguments
cannot be used simultaneously.}

\item{cursor}{A logical indicating whether to use cursor-based paging when
querying the OpenAIRE Graph API (\code{TRUE}) or not (\code{FALSE}). Note that the
\code{cursor} (cursor-based paging) and \code{page} (offset-based paging) arguments
cannot be used simultaneously.}
}
\value{
A list of class \code{oag_options} with the options requirements to pass
to \code{oa_query()}. Note that at this stage, the validity of the options has not
been checked yet.
}
\description{
This function collects the options requirements to use when building a query.
The validity of the options are collected without any checks. The checks will
take place later, after being passed to \code{oag_query()}.
}
\examples{
# Example of query options with offset-based paging with sorting
oag_options(
  sortBy = c(prevalence = "ASC"),
  includeStats = TRUE,
  pageSize = 50,
  page = 1
)

# Example of query options with cursor-based paging and sorting
oag_options(
  sortBy = c(prevalence = "DESC"),
  includeStats = FALSE,
  pageSize = 100,
  cursor = TRUE
)
}
