% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oag_parser.R
\name{oag_parse_object}
\alias{oag_parse_object}
\title{Parse the research products objects}
\usage{
oag_parse_object(
  object,
  entity = NULL,
  selection = NULL,
  type = c("publication", "dataset", "software", "other")
)
}
\arguments{
\item{object}{An OpenAIRE Graph object as returned by \code{oag_fetch()} or
\code{oag_request()}.}

\item{entity}{A string with the entity to query. It must match an entity
returned by \code{oag_entities()}.}

\item{selection}{A character vector with the variables to select in the
OpenAIRE Graph object passed to \code{object}.}

\item{type}{A string with the type of research product contained in \code{object}}
}
\value{
A tibble with the parsed object.
}
\description{
Parse the research products objects
}
\examples{
\dontrun{
# Fetch some "research products" data from the OpenAIRE Graph
res_prod <-  oag_fetch(
  "research-products",
  type = "publication",
  relProjectFundingShortName = "SNSF",
  fromPublicationDate = "2023-01-01",
  toPublicationDate = "2023-01-31",
  options = oag_options(pageSize = 100, cursor = TRUE)
)

# Parsing the returned research products object
res_prod_df <- oag_parse_object(
  res_prod,
  entity = "research-products",
  type = "publication"
)

# Fetch some "organizations" data from the OpenAIRE Graph
res_org <- oag_fetch(
  "organizations",
  countryCode = "CH",
  options = oag_options(pageSize = 100, cursor = TRUE)
)

# Parsing the returned organizations object
res_org_df <- oag_parse_object(res_org, entity = "organizations")

# Fetch some "projects" data from the OpenAIRE Graph
res_proj <- oag_fetch(
  "projects",
  fundingShortName = "SNSF",
  startYear = "2025",
  options = oag_options(pageSize = 100, cursor = TRUE)
)

# Parsing the returned projects object
res_proj_df <- oag_parse_object(res_proj, entity = "projects")

# Fetch some "persons" data from the OpenAIRE Graph
res_prsn <- oag_fetch(
  "persons",
  lastName = "Gorin",
  options = oag_options(pageSize = 100, cursor = TRUE)
)

# Parsing the returned projects object
res_prsn_df <- oag_parse_object(res_prsn, entity = "persons")

# Fetch some "persons" data from the OpenAIRE Graph
res_ds <- oag_fetch(
  "datasources",
  country = "CH",
  options = oag_options(pageSize = 100, cursor = TRUE)
)

# Parsing the returned projects object
res_ds_df <- oag_parse_object(res_ds, entity = "datasources")
}
}
