% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oag_fetch.R
\name{oag_query}
\alias{oag_query}
\title{Compose a query to the OpenAIRE Graph API}
\usage{
oag_query(entity, ..., options = NULL)
}
\arguments{
\item{entity}{A string with the entity to query. It must match an entity
returned by \code{oag_entities()}.}

\item{...}{Filter(s) to use when composing the query. The filter(s) must be
named and their value must be a string. The name of the filter must be a
valid filter referenced in the OpenAIRE Graph API documentation. For example,
\code{publicationYear = "2020"} can be used to filter publication from 2020 when
working with "research-products" entity.}

\item{options}{A set of options (created with \code{opa_options()}) used to
compose the query. Options allow to control sorting, paging, etc. See
\code{?opa_options()} for more details.}
}
\value{
A string with the query URL.
}
\description{
The OpenAIRE Graph API works with filters and options (see
\url{https://graph.openaire.eu/docs/apis/graph-api/}). The function
interprets additional (named) argument(s) passed to the \code{...} argument as
filter(s). The filter(s) must be named with the name of the field to filter,
as described in the OpenAIRE Graph API. Options can be specified with
\code{opa_options()}.
}
\examples{
oag_query(
  "research-products",
  publicationYear = "2020",
  options = oag_options(sortBy = c(relevance = "ASC"))
)
}
