% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oag_options.R
\name{oag_set_options}
\alias{oag_set_options}
\title{Check and format all options from an object of class \code{oag_options}}
\usage{
oag_set_options(
  entity,
  sortBy = NULL,
  includeStats = NULL,
  pageSize = NULL,
  page = NULL,
  cursor = NULL,
  is_cursor_next = FALSE,
  call = rlang::caller_env()
)
}
\arguments{
\item{entity}{A string with the entity to query. It must match an entity
returned by \code{oag_entities()}.}

\item{sortBy}{A named character vector with the field(s) on which to sort the
results. Names must be valid sorting fields, only accepting the values "ASC"
(ascending) and "DESC" (descending) as input. Note that the fields on which
results can be sorted vary across entities. To get the list of sorting fields
for a given entity use \code{get_sorting_fields()}.}

\item{includeStats}{A scalar logical indicating whether statistics about the
query should be included (\code{TRUE}) or not (\code{FALSE}) in the results.}

\item{pageSize}{A positive integer indicating the number of records per page
a query should return. The OpenAIRE Graph API has a limit of 100 records per
page.}

\item{page}{A positive integer indicating the specific page the query should
return. When \code{page} is kept NULL, this triggers an offset-based paging query.
Note that the \code{page} (offset-based paging) and \code{cursor} (cursor-based paging)
arguments cannot be used simultaneously.}

\item{cursor}{A logical indicating whether to use cursor-based paging when
querying the OpenAIRE Graph API (\code{TRUE}) or not (\code{FALSE}). Note that the
\code{cursor} (cursor-based paging) and \code{page} (offset-based paging) arguments
cannot be used simultaneously.}

\item{is_cursor_next}{\strong{Do not use this argument}. It is used internally by
\code{oag_fetch()} to indicate whether the \code{cursor} argument must be interpreted
as the method to use ("cursor-based paging") or as the next cursor.}

\item{call}{The caller environment passed to \code{cli::cli_abort()}.}
}
\value{
A string with options requirements formatted according to the
OpenAIRE Graph API's documentation.
}
\description{
Check and format all options from an object of class \code{oag_options}
}
\keyword{internal}
