% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oag_parser.R
\name{parse_research_products}
\alias{parse_research_products}
\alias{parse_entity_organizations}
\alias{parse_entity_projects}
\title{Parse the research products objects}
\usage{
parse_research_products(
  object,
  type = c("publication", "dataset", "software", "other"),
  selection = NULL
)

parse_entity_organizations(object, selection = NULL)

parse_entity_projects(object, selection = NULL)
}
\arguments{
\item{object}{An OpenAIRE Graph object as returned by \code{oag_fetch()} or
\code{oag_request()}.}

\item{type}{A string with the type of research product contained in \code{object}}

\item{selection}{A character vector with the variables to select in the
OpenAIRE Graph object passed to \code{object}.}
}
\value{
A tibble with the parsed object.
}
\description{
Parse the research products objects
}
\examples{
\dontrun{
# Fetch some "research products" data from the OpenAIRE Graph
res_prod <-  oag_fetch(
  "research-products",
  type = "publication",
  relProjectFundingShortName = "SNSF",
  fromPublicationDate = "2023-01-01",
  toPublicationDate = "2023-01-31",
  options = oag_options(pageSize = 100, cursor = TRUE)
)

# Parsing the returned research products object
res_prod_df <- parse_research_products(res_prod, type = "publication")

# Fetch some "organizations" data from the OpenAIRE Graph
res_org <- oag_fetch(
  "organizations",
  countryCode = "CH",
  options = oag_options(pageSize = 100, cursor = TRUE)
)

# Parsing the returned organizations object
res_org_df <- parse_entity_organizations(res_org)

# Fetch some "projects" data from the OpenAIRE Graph
res_proj <- oag_fetch(
  "projects",
  fundingShortName = "SNSF",
  startYear = "2025",
  options = oag_options(pageSize = 100, cursor = TRUE)
)

# Parsing the returned projects object
res_proj_df <- parse_entity_projects(res_proj)
}
}
