% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oag_parser.R
\name{parse_research_products}
\alias{parse_research_products}
\title{Parse the research products objects}
\usage{
parse_research_products(
  prod,
  type = c("publication", "dataset", "software", "other"),
  selection = NULL
)
}
\arguments{
\item{prod}{A list with research products objects as returned by
\code{oag_fetch()} or \code{oag_request()}.}

\item{type}{A string with the type of research product contained in \code{prod}}

\item{selection}{A character vector with the variables to select in the
research product object passed to \code{prod}.}
}
\value{
A tibble with the parsed research products.
}
\description{
Parse the research products objects
}
\examples{
\dontrun{
# Fetch some data from the OpenAIRE Graph
res_prod <-  oag_fetch(
  "research-products",
  type = "publication",
  relProjectFundingShortName = "SNSF",
  fromPublicationDate = "2023-01-01",
  toPublicationDate = "2023-01-31",
  options = oag_options(pageSize = 100, cursor = TRUE)
)

# Parsing the returned research products objects
res_prod_df <- parse_research_products(res_prod, type = "publication")
}
}
