% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oag_options.R
\name{sorting_error_msg}
\alias{sorting_error_msg}
\title{Compose an error message for sorting option issues with pre-formatted help}
\usage{
sorting_error_msg(msg, .call = rlang::current_env())
}
\arguments{
\item{msg}{A string with the message to add to the pre-formatted error
message.}

\item{call}{The caller environment passed to \code{cli::cli_abort()}.}
}
\value{
Nothing, use for side-effect only.
}
\description{
Compose an error message for sorting option issues with pre-formatted help
}
